+++ src/main/resources/sql/postgresql/jds_ref_entity_enum.sql
@@ -1,6 +1,6 @@
 CREATE TABLE jds_ref_entity_enum (
-  entity_id BIGINT,
-  field_id  BIGINT,
+  entity_id INTEGER,
+  field_id  INTEGER,
   PRIMARY KEY (entity_id, field_id),
   FOREIGN KEY (entity_id) REFERENCES jds_ref_entity (id) ON DELETE CASCADE,
   FOREIGN KEY (field_id) REFERENCES jds_ref_field (id) ON DELETE CASCADE
